Code fonctionnant avec la mémoire

+++ projet/app/ecrire.h
@@ -7,5 +7,5 @@
 
 void ecrire_memoire(Memoire24CXXX& memoire, uint8_t pourcentageMoteurG, uint8_t pourcentageMoteurD, volatile uint16_t& addresse);
 void indiquer_fin_memoire(Memoire24CXXX& memoire, volatile uint16_t& addresse);
-uint8_t combine(uint16_t pourcentageMoteurG, uint16_t pourcentageMoteurD);
+uint8_t combine(uint8_t pourcentageMoteurG, uint8_t pourcentageMoteurD);
 void dechiffrer_donnee(uint8_t donnee, uint8_t& pourcentageMoteurG, uint8_t& pourcentageMoteurD);
